+++ src/buildings/model/buildings-channel-condition-model.cc
@@ -99,6 +99,35 @@ BuildingsChannelConditionModel::GetChannelCondition(Ptr<const MobilityModel> a,
         {
             NS_LOG_DEBUG("a and b are indoor in different buildings");
             cond->SetLosCondition(ChannelCondition::LosConditionValue::NLOS);
+
+            ChannelCondition::O2iLowHighConditionValue lowHighLossConditionA1;
+            ChannelCondition::O2iLowHighConditionValue lowHighLossConditionB1;
+
+            // Low losses considered for Wood or ConcreteWithWindows, while
+            // high losses for ConcreteWithoutWindows and StoneBlocks
+            lowHighLossConditionA1 =
+                a1->GetBuilding()->GetExtWallsType() == Building::ExtWallsType_t::Wood ||
+                        a1->GetBuilding()->GetExtWallsType() ==
+                            Building::ExtWallsType_t::ConcreteWithWindows
+                    ? ChannelCondition::O2iLowHighConditionValue::LOW
+                    : ChannelCondition::O2iLowHighConditionValue::HIGH;
+
+            lowHighLossConditionB1 =
+                b1->GetBuilding()->GetExtWallsType() == Building::ExtWallsType_t::Wood ||
+                        b1->GetBuilding()->GetExtWallsType() ==
+                            Building::ExtWallsType_t::ConcreteWithWindows
+                    ? ChannelCondition::O2iLowHighConditionValue::LOW
+                    : ChannelCondition::O2iLowHighConditionValue::HIGH;
+
+            if (lowHighLossConditionA1 == ChannelCondition::O2iLowHighConditionValue::HIGH ||
+                lowHighLossConditionB1 == ChannelCondition::O2iLowHighConditionValue::HIGH)
+            {
+                cond->SetO2iLowHighCondition(ChannelCondition::O2iLowHighConditionValue::HIGH);
+            }
+            else
+            {
+                cond->SetO2iLowHighCondition(ChannelCondition::O2iLowHighConditionValue::LOW);
+            }
         }
     }
     else // outdoor to indoor case
@@ -107,6 +136,31 @@ BuildingsChannelConditionModel::GetChannelCondition(Ptr<const MobilityModel> a,
 
         NS_LOG_DEBUG("a is indoor and b outdoor or vice-versa");
         cond->SetLosCondition(ChannelCondition::LosConditionValue::NLOS);
+
+        ChannelCondition::O2iLowHighConditionValue lowHighLossCondition;
+        if (isAIndoor)
+        {
+            // Low losses considered for Wood or ConcreteWithWindows, while
+            // high losses for ConcreteWithoutWindows and StoneBlocks
+            lowHighLossCondition =
+                a1->GetBuilding()->GetExtWallsType() == Building::ExtWallsType_t::Wood ||
+                        a1->GetBuilding()->GetExtWallsType() ==
+                            Building::ExtWallsType_t::ConcreteWithWindows
+                    ? ChannelCondition::O2iLowHighConditionValue::LOW
+                    : ChannelCondition::O2iLowHighConditionValue::HIGH;
+
+            cond->SetO2iLowHighCondition(lowHighLossCondition);
+        }
+        else
+        {
+            lowHighLossCondition =
+                b1->GetBuilding()->GetExtWallsType() == Building::ExtWallsType_t::Wood ||
+                        b1->GetBuilding()->GetExtWallsType() ==
+                            Building::ExtWallsType_t::ConcreteWithWindows
+                    ? ChannelCondition::O2iLowHighConditionValue::LOW
+                    : ChannelCondition::O2iLowHighConditionValue::HIGH;
+            cond->SetO2iLowHighCondition(lowHighLossCondition);
+        }
     }
 
     return cond;