+++ CHANGES.md
@@ -65,6 +65,7 @@ Changes from ns-3.38 to ns-3-dev
 
 ### Changed behavior
 
+* (buildings) Calculation of the O2I Low/High Building Penetration Losses based on 3GPP 38.901 7.4.3.1 was missing. These losses are now included in the pathloss calculation when buildings are present.
 * (network) The function `Buffer::Allocate` will over-provision `ALLOC_OVER_PROVISION` bytes when allocating buffers for packets. `ALLOC_OVER_PROVISION` is currently set to 100 bytes.
 
 Changes from ns-3.37 to ns-3.38